+++ BankService/Deal/src/test/java/com/munsun/deal/services/clients/CalculatorFeignClientUnitTests.java
@@ -7,6 +7,8 @@ import com.maciejwalkowiak.wiremock.spring.ConfigureWireMock;
 import com.maciejwalkowiak.wiremock.spring.EnableWireMock;
 import com.maciejwalkowiak.wiremock.spring.InjectWireMock;
 import com.munsun.deal.dto.*;
+import com.munsun.deal.exceptions.PrescoringException;
+import com.munsun.deal.exceptions.ScoringException;
 import com.munsun.deal.utils.TestUtils;
 import com.munsun.deal.services.impl.clients.CalculatorFeignClient;
 import feign.FeignException;
@@ -27,10 +29,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @SpringBootTest
 @EnableWireMock(
-        @ConfigureWireMock(name="calculator-client", property = "${client.calculator.url}")
+        @ConfigureWireMock(name="calculator", property = "${clients.calculator.name}")
 )
 public class CalculatorFeignClientUnitTests {
-    @InjectWireMock("calculator-client")
+    @InjectWireMock("calculator")
     private WireMockServer server;
     @Autowired
     private ObjectMapper mapper;
@@ -42,7 +44,7 @@ public class CalculatorFeignClientUnitTests {
     public void givenRequestWithLoanStatementRequest_whenSendRequestToServer_thenReturnResponseListLoanOfferStatus200() throws JsonProcessingException {
         LoanStatementRequestDto loanRequest = TestUtils.getLoanStatementRequestDto();
         List<LoanOfferDto> expectedLoanOffers = TestUtils.getAnnuitentPaymentListLoanOffersDtoAmount10_000Term12();
-        server.stubFor(post(LOAN_OFFERS_ENDPOINT_CALCULATOR+"?typePayments=ANNUITY")
+        server.stubFor(post(LOAN_OFFERS_ENDPOINT_CALCULATOR+"?typePayment=ANNUITY")
                 .willReturn(aResponse()
                     .withHeader("Content-Type", "application/json")
                     .withBody(mapper.writeValueAsString(expectedLoanOffers))));
@@ -59,7 +61,7 @@ public class CalculatorFeignClientUnitTests {
     public void givenRequestWithScoringDataDto_whenSendRequestToServer_thenReturnResponseWithCreditDtoStatus200() throws JsonProcessingException {
         ScoringDataDto scoringData = TestUtils.getScoringDataDto();
         CreditDto expectedCreditDto = TestUtils.getCreditDto();
-        server.stubFor(post(CALC_CREDIT_ENDPOINT_CALCULATOR+"?typePayments=ANNUITY")
+        server.stubFor(post(CALC_CREDIT_ENDPOINT_CALCULATOR+"?typePayment=ANNUITY")
                 .willReturn(aResponse()
                         .withHeader("Content-Type", "application/json")
                         .withBody(mapper.writeValueAsString(expectedCreditDto))));
@@ -77,20 +79,15 @@ public class CalculatorFeignClientUnitTests {
     public void givenRequestWithLoanRequestInvalid_whenSendRequest_thenReturnResponseWithErrorMessageStatus400() throws JsonProcessingException {
         LoanStatementRequestDto loanRequest = TestUtils.getLoanStatementRequestDtoInvalidAmount();
         ErrorMessageDto errorMessage = TestUtils.getErrorMessageInvalidAmount();
-        server.stubFor(post(LOAN_OFFERS_ENDPOINT_CALCULATOR+"?typePayments=ANNUITY")
+        server.stubFor(post(LOAN_OFFERS_ENDPOINT_CALCULATOR+"?typePayment=ANNUITY")
                 .willReturn(aResponse()
                         .withStatus(HttpStatus.BAD_REQUEST.value())
                         .withHeader("Content-Type", "application/json")
                         .withBody(mapper.writeValueAsString(errorMessage))));
 
-        try {
-            client.getLoanOffers(TypePayments.ANNUITY, loanRequest);
-        } catch (FeignException e) {
-            assertThat(e.contentUTF8())
-                    .contains(errorMessage.getMessage());
-            assertThat(e.status())
-                    .isEqualTo(HttpStatus.BAD_REQUEST.value());
-        }
+        assertThatThrownBy(() ->
+            client.getLoanOffers(TypePayments.ANNUITY, loanRequest)
+        ).isInstanceOf(PrescoringException.class);
     }
 
     @DisplayName("Test send request with ScoringDataDto, scoring error")
@@ -98,19 +95,14 @@ public class CalculatorFeignClientUnitTests {
     public void givenRequestWithScoringDataDtoInvalid_whenSendRequest_thenReturnResponseWithErrorMessageStatus500() throws JsonProcessingException {
         LoanStatementRequestDto loanRequest = TestUtils.getLoanStatementRequestDtoInvalidAmount();
         ErrorMessageDto errorMessage = TestUtils.getErrorMessageScoringError();
-        server.stubFor(post(LOAN_OFFERS_ENDPOINT_CALCULATOR+"?typePayments=ANNUITY")
+        server.stubFor(post(LOAN_OFFERS_ENDPOINT_CALCULATOR+"?typePayment=ANNUITY")
                 .willReturn(aResponse()
                         .withStatus(HttpStatus.INTERNAL_SERVER_ERROR.value())
                         .withHeader("Content-Type", "application/json")
                         .withBody(mapper.writeValueAsString(errorMessage))));
 
-        try {
-            client.getLoanOffers(TypePayments.ANNUITY, loanRequest);
-        } catch (FeignException e) {
-            assertThat(e.contentUTF8())
-                    .contains(errorMessage.getMessage());
-            assertThat(e.status())
-                    .isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
-        }
+        assertThatThrownBy(() ->
+                client.getLoanOffers(TypePayments.ANNUITY, loanRequest)
+        ).isInstanceOf(ScoringException.class);
     }
 }