+++ BankService/Deal/changelog.md
@@ -3,6 +3,8 @@
 ### Patch:
 
 - удалены перехватчики для логирования запросов/ответов;
+- добавлены аспекты с счетчиками для подсчета заявок со статусом APPROVED и CREDIT_ISSUED. [Статья](https://habr.com/ru/companies/otus/articles/650871/), [Статья](https://docs.micrometer.io/micrometer/reference/concepts/counters.html);
+- добавлен ErrorDecoder для feign client.
 
 
 # 1.1.2 - 2025.03.03